# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marketplace de Servicios API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Cannot delete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/listings/{id}:
    delete:
      operationId: adminDeleteListing
      tags: [admin]
      summary: Delete any listing (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/listings/{id}/approve:
    patch:
      operationId: adminApproveListing
      tags: [admin]
      summary: Approve or reject a listing (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                approved:
                  type: boolean
              required:
                - approved
      responses:
        "200":
          description: Listing approval status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/jobs/{id}/approve:
    patch:
      operationId: adminApproveJob
      tags: [admin]
      summary: Approve or reject a job posting (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                approved:
                  type: boolean
              required:
                - approved
      responses:
        "200":
          description: Job approval status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/categories:
    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Create a category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/categories/{id}:
    patch:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Update a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      summary: Delete a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Category in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [provider, client, admin, company]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        notifyEmail:
          type: boolean
        companyName:
          type: ["string", "null"]
        cuit:
          type: ["string", "null"]
        companyAddress:
          type: ["string", "null"]
        companyIndustry:
          type: ["string", "null"]
        companyApproved:
          type: boolean
        cvUrl:
          type: ["string", "null"]
        cvPublic:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - notifyEmail
        - companyApproved
        - cvPublic
        - createdAt
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: ["string", "null"]
        type:
          type: string
          enum: [service, product]
        description:
          type: ["string", "null"]
      required:
        - id
        - name
        - type
...
    AdminUpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
          enum: [provider, client, admin, null]
        phone:
          type: ["string", "null"]
        locality:
          type: ["string", "null"]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        icon:
          type: ["string", "null"]
        type:
          type: string
          enum: [service, product]
        description:
          type: ["string", "null"]
      required:
        - name
        - type
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
          enum: [service, product, null]
        description:
          type: ["string", "null"]
...