# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketplace de Servicios API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  locality?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  notifyEmail: boolean;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  cuit?: string | null;
  /** @nullable */
  companyAddress?: string | null;
  /** @nullable */
  companyIndustry?: string | null;
  companyApproved: boolean;
  /** @nullable */
  cvUrl?: string | null;
  cvPublic: boolean;
  createdAt: string;
}